Moving the _2s into their main ones
