--- enterprise_extensions/models_2.py
@@ -1,2182 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import, division, print_function, unicode_literals
-import numpy as np
-from collections import OrderedDict
-
-from enterprise.signals import parameter
-from enterprise.signals import selections
-from enterprise.signals import signal_base
-from enterprise.signals import white_signals
-from enterprise.signals import gp_signals
-from enterprise.signals import deterministic_signals
-from enterprise import constants as const
-
-from enterprise_extensions import model_utils
-from enterprise_extensions import deterministic
-
-from enterprise_extensions.timing_2 import (timing_block,filter_Mmat)
-from enterprise_extensions.blocks import (
-    white_noise_block,
-    red_noise_block,
-    dm_noise_block,
-    chromatic_noise_block,
-    common_red_noise_block,
-)
-from enterprise_extensions.chromatic.solar_wind import solar_wind_block
-from enterprise_extensions import chromatic as chrom
-from enterprise_extensions import dropout as do
-
-"""
-PTA models from paper
-"""
-
-def model_singlepsr_noise(
-    psr,
-    tm_var=False,
-    tm_linear=False,
-    tm_param_list=[],
-    tm_param_dict={},
-    tm_prior="uniform",
-    red_var=True,
-    psd="powerlaw",
-    red_select=None,
-    noisedict=None,
-    tm_svd=False,
-    tm_norm=True,
-    white_vary=True,
-    components=30,
-    upper_limit=False,
-    wideband=False,
-    gamma_val=None,
-    dm_var=False,
-    dm_type="gp",
-    dmgp_kernel="diag",
-    dm_psd="powerlaw",
-    dm_nondiag_kernel="periodic",
-    dmx_data=None,
-    dm_annual=False,
-    gamma_dm_val=None,
-    chrom_gp=False,
-    chrom_gp_kernel="nondiag",
-    chrom_psd="powerlaw",
-    chrom_idx=4,
-    chrom_kernel="periodic",
-    dm_expdip=False,
-    dmexp_sign="negative",
-    dm_expdip_idx=2,
-    dm_expdip_tmin=None,
-    dm_expdip_tmax=None,
-    num_dmdips=1,
-    dmdip_seqname=None,
-    dm_cusp=False,
-    dm_cusp_sign="negative",
-    dm_cusp_idx=2,
-    dm_cusp_sym=False,
-    dm_cusp_tmin=None,
-    dm_cusp_tmax=None,
-    num_dm_cusps=1,
-    dm_cusp_seqname=None,
-    dm_dual_cusp=False,
-    dm_dual_cusp_tmin=None,
-    dm_dual_cusp_tmax=None,
-    dm_dual_cusp_sym=False,
-    dm_dual_cusp_idx1=2,
-    dm_dual_cusp_idx2=4,
-    dm_dual_cusp_sign="negative",
-    num_dm_dual_cusps=1,
-    dm_dual_cusp_seqname=None,
-    dm_sw_deter=False,
-    dm_sw_gp=False,
-    swgp_prior=None,
-    swgp_basis=None,
-    coefficients=False,
-    extra_sigs=None,
-):
-    """
-    Single pulsar noise model
-    :param psr: enterprise pulsar object
-    :param tm_var: explicitly vary the timing model parameters
-    :param tm_linear: vary the timing model in the linear approximation
-    :param tm_param_list: an explicit list of timing model parameters to vary
-    :param tm_param_dict: a nested dictionary of parameters to vary in the model and their user defined values and priors
-    :param tm_prior: prior type on varied timing model parameters {'Uniform','bounded-normal'}
-    :param red var: include red noise in the model
-    :param psd: red noise psd model
-    :param noisedict: dictionary of noise parameters
-    :param tm_svd: boolean for svd-stabilised timing model design matrix
-    :param tm_norm: normalize the timing model, or provide custom normalization
-    :param white_vary: boolean for varying white noise or keeping fixed
-    :param components: number of modes in Fourier domain processes
-    :param upper_limit: whether to do an upper-limit analysis
-    :param wideband: whether to include ecorr in the white noise model
-    :param gamma_val: red noise spectral index to fix
-    :param dm_var: whether to explicitly model DM-variations
-    :param dm_type: gaussian process ('gp') or dmx ('dmx')
-    :param dmgp_kernel: diagonal in frequency or non-diagonal
-    :param dm_psd: power-spectral density of DM variations
-    :param dm_nondiag_kernel: type of time-domain DM GP kernel
-    :param dmx_data: supply the DMX data from par files
-    :param dm_annual: include an annual DM signal
-    :param gamma_dm_val: spectral index of power-law DM variations
-    :param chrom_gp: include general chromatic noise
-    :param chrom_gp_kernel: GP kernel type to use in chrom ['diag','nondiag']
-    :param chrom_psd: power-spectral density of chromatic noise
-    :param chrom_idx: frequency scaling of chromatic noise
-    :param chrom_kernel: Type of 'nondiag' kernel to use in chrom GP.
-    :param dm_expdip: inclue a DM exponential dip
-    :param dmexp_sign: set the sign parameter for dip
-    :param dm_expdip_idx: chromatic index of exponential dip
-    :param dm_expdip_tmin: sampling minimum of DM dip epoch
-    :param dm_expdip_tmax: sampling maximum of DM dip epoch
-    :param num_dmdips: number of dm exponential dips
-    :param dmdip_seqname: name of dip sequence
-    :param dm_cusp: inclue a DM exponential cusp
-    :param dm_cusp_sign: set the sign parameter for cusp
-    :param dm_cusp_idx: chromatic index of exponential cusp
-    :param dm_cusp_tmin: sampling minimum of DM cusp epoch
-    :param dm_cusp_tmax: sampling maximum of DM cusp epoch
-    :param dm_cusp_sym: make exponential cusp symmetric
-    :param num_dm_cusps: number of dm exponential cusps
-    :param dm_cusp_seqname: name of cusp sequence
-    :param dm_dual_cusp: include a DM cusp with two chromatic indices
-    :param dm_dual_cusp_tmin: sampling minimum of DM dual cusp epoch
-    :param dm_dual_cusp_tmax: sampling maximum of DM dual cusp epoch
-    :param dm_dual_cusp_idx1: first chromatic index of DM dual cusp
-    :param dm_dual_cusp_idx2: second chromatic index of DM dual cusp
-    :param dm_dual_cusp_sym: make dual cusp symmetric
-    :param dm_dual_cusp_sign: set the sign parameter for dual cusp
-    :param num_dm_dual_cusps: number of DM dual cusps
-    :param dm_dual_cusp_seqname: name of dual cusp sequence
-    :param dm_scattering: whether to explicitly model DM scattering variations
-    :param chrom_kernel: type of time-domain DM GP kernel for the scattering
-        variations
-    :param dm_sw_deter: use the deterministic solar wind model
-    :param dm_sw_gp: add a Gaussian process perturbation to the deterministic
-        solar wind model.
-    :param swgp_prior: prior is currently set automatically
-    :param swgp_basis: ['powerlaw', 'periodid', 'sq_exp']
-    :param coefficients: explicitly include latent coefficients in model
-    :param extra_sigs: Any additional `enterprise` signals to be added to the
-        model.
-
-    :return s: single pulsar noise model
-    """
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # timing model
-    if not tm_var:
-        s = gp_signals.TimingModel(
-            use_svd=tm_svd, normed=tm_norm, coefficients=coefficients
-        )
-    else:
-        # create new attribute for enterprise pulsar object
-        psr.tm_params_orig = OrderedDict.fromkeys(psr.t2pulsar.pars())
-        for key in psr.tm_params_orig:
-            psr.tm_params_orig[key] = (psr.t2pulsar[key].val, psr.t2pulsar[key].err)
-        if not tm_linear:
-            s = timing_block(
-                tm_param_list=tm_param_list,
-                prior_type=tm_prior,
-                prior_sigma=2.0,
-                prior_lower_bound=-3.0,
-                prior_upper_bound=3.0,
-                tm_param_dict=tm_param_dict,
-            )
-        else:
-            pass
-
-    # red noise
-    if red_var:
-        s += red_noise_block(
-            psd=psd,
-            prior=amp_prior,
-            components=components,
-            gamma_val=gamma_val,
-            coefficients=coefficients,
-            select=red_select,
-        )
-
-    # DM variations
-    if dm_var:
-        if dm_type == "gp":
-            if dmgp_kernel == "diag":
-                s += dm_noise_block(
-                    gp_kernel=dmgp_kernel,
-                    psd=dm_psd,
-                    prior=amp_prior,
-                    components=components,
-                    gamma_val=gamma_dm_val,
-                    coefficients=coefficients,
-                )
-            elif dmgp_kernel == "nondiag":
-                s += dm_noise_block(
-                    gp_kernel=dmgp_kernel,
-                    nondiag_kernel=dm_nondiag_kernel,
-                    coefficients=coefficients,
-                )
-        elif dm_type == "dmx":
-            s += chrom.dmx_signal(dmx_data=dmx_data[psr.name])
-        if dm_annual:
-            s += chrom.dm_annual_signal()
-        if chrom_gp:
-            s += chromatic_noise_block(
-                gp_kernel=chrom_gp_kernel,
-                psd=chrom_psd,
-                idx=chrom_idx,
-                components=components,
-                nondiag_kernel=chrom_kernel,
-                coefficients=coefficients,
-            )
-
-        if dm_expdip:
-            if dm_expdip_tmin is None and dm_expdip_tmax is None:
-                tmin = [psr.toas.min() / 86400 for ii in range(num_dmdips)]
-                tmax = [psr.toas.max() / 86400 for ii in range(num_dmdips)]
-            else:
-                tmin = (
-                    dm_expdip_tmin
-                    if isinstance(dm_expdip_tmin, list)
-                    else [dm_expdip_tmin]
-                )
-                tmax = (
-                    dm_expdip_tmax
-                    if isinstance(dm_expdip_tmax, list)
-                    else [dm_expdip_tmax]
-                )
-            if dmdip_seqname is not None:
-                dmdipname_base = "dmexp_" + dmdip_seqname + "_"
-            else:
-                dmdipname_base = "dmexp_"
-            dm_expdip_idx = (
-                dm_expdip_idx if isinstance(dm_expdip_idx, list) else [dm_expdip_idx]
-            )
-            for dd in range(1, num_dmdips + 1):
-                s += chrom.dm_exponential_dip(
-                    tmin=tmin[dd - 1],
-                    tmax=tmax[dd - 1],
-                    idx=dm_expdip_idx[dd - 1],
-                    sign=dmexp_sign,
-                    name=dmdipname_base + str(dd),
-                )
-        if dm_cusp:
-            if dm_cusp_tmin is None and dm_cusp_tmax is None:
-                tmin = [psr.toas.min() / 86400 for ii in range(num_dm_cusps)]
-                tmax = [psr.toas.max() / 86400 for ii in range(num_dm_cusps)]
-            else:
-                tmin = (
-                    dm_cusp_tmin if isinstance(dm_cusp_tmin, list) else [dm_cusp_tmin]
-                )
-                tmax = (
-                    dm_cusp_tmax if isinstance(dm_cusp_tmmax, list) else [dm_cusp_tmax]
-                )
-            if dm_cusp_seqname is not None:
-                cusp_name_base = "dm_cusp_" + dm_cusp_seqname + "_"
-            else:
-                cusp_name_base = "dm_cusp_"
-            dm_cusp_idx = (
-                dm_cusp_idx if isinstance(dm_cusp_idx, list) else [dm_cusp_idx]
-            )
-            for dd in range(1, num_dm_cusps + 1):
-                s += chrom.dm_exponential_cusp(
-                    tmin=tmin[dd - 1],
-                    tmax=tmax[dd - 1],
-                    idx=dm_cusp_idx,
-                    sign=dm_cusp_sign,
-                    symmetric=dm_cusp_sym,
-                    name=cusp_name_base + str(dd),
-                )
-        if dm_dual_cusp:
-            if dm_dual_cusp_tmin is None and dm_cusp_tmax is None:
-                tmin = psr.toas.min() / 86400
-                tmax = psr.toas.max() / 86400
-            else:
-                tmin = dm_dual_cusp_tmin
-                tmax = dm_dual_cusp_tmax
-            if dm_dual_cusp_seqname is not None:
-                dual_cusp_name_base = "dm_dual_cusp_" + dm_cusp_seqname + "_"
-            else:
-                dual_cusp_name_base = "dm_dual_cusp_"
-            for dd in range(1, num_dm_dual_cusps + 1):
-                s += chrom.dm_dual_exp_cusp(
-                    tmin=tmin,
-                    tmax=tmax,
-                    idx1=dm_dual_cusp_idx1,
-                    idx2=dm_dual_cusp_idx2,
-                    sign=dm_dual_cusp_sign,
-                    symmetric=dm_dual_cusp_sym,
-                    name=dual_cusp_name_base + str(dd),
-                )
-        if dm_sw_deter:
-            Tspan = psr.toas.max() - psr.toas.min()
-            s += solar_wind_block(
-                ACE_prior=True,
-                include_swgp=dm_sw_gp,
-                swgp_prior=swgp_prior,
-                swgp_basis=swgp_basis,
-                Tspan=Tspan,
-            )
-
-    if extra_sigs is not None:
-        s += extra_sigs
-    # adding white-noise, and acting on psr objects
-    if "NANOGrav" in psr.flags["pta"] and not wideband:
-        s2 = s + white_noise_block(vary=white_vary, inc_ecorr=True)
-        model = s2(psr)
-    else:
-        s3 = s + white_noise_block(vary=white_vary, inc_ecorr=False)
-        model = s3(psr)
-
-    # set up PTA
-    pta = signal_base.PTA([model])
-
-    # set white noise parameters
-    if not white_vary:
-        if noisedict is None:
-            print("No noise dictionary provided!...")
-        else:
-            noisedict = noisedict
-            pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_1(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    upper_limit=False,
-    bayesephem=False,
-    wideband=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with only white and red noise:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1. Optional physical ephemeris modeling.
-
-
-    :param psd:
-        Choice of PSD function [e.g. powerlaw (default), turnover, tprocess]
-    :param noisedict:
-        Dictionary of pulsar noise properties. Can provide manually,
-        or the code will attempt to find it.
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    :param wideband:
-        Use wideband par and tim files. Ignore ECORR. Set to False by default.
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(psd=psd, prior=amp_prior, Tspan=Tspan, components=components)
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_2a(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    bayesephem=False,
-    be_type="orbel",
-    wideband=False,
-    select="backend",
-    pshift=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 2A from the analysis paper:
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-    global:
-        1.Common red noise modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        2. Optional physical ephemeris modeling.
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum']. 'powerlaw' is default
-        value.
-    :param noisedict:
-        Dictionary of pulsar noise properties. Can provide manually,
-        or the code will attempt to find it.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    :param be_type:
-        orbel, orbel-v2, setIII
-    :param wideband:
-        Use wideband par and tim files. Ignore ECORR. Set to False by default.
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(prior=amp_prior, Tspan=Tspan, components=components)
-
-    # common red noise block
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        name="gw",
-        pshift=pshift,
-    )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(
-            use_epoch_toas=True, model=be_type
-        )
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True, select=select)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False, select=select)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_general(
-    psrs,
-    tm_var=False,
-    tm_linear=False,
-    tm_param_list=[],
-    ltm_exclude_list=[],
-    exclude = True,
-    tm_param_dict={},
-    tm_prior="uniform",
-    nltm_plus_ltm = False,
-    common_var=True,
-    common_psd="powerlaw",
-    red_psd="powerlaw",
-    orf=None,
-    common_components=30,
-    red_components=30,
-    dm_components=30,
-    modes=None,
-    wgts=None,
-    logfreq=False,
-    nmodes_log=10,
-    noisedict=None,
-    tm_svd=False,
-    tm_norm=True,
-    gamma_common=None,
-    upper_limit=False,
-    upper_limit_red=None,
-    upper_limit_dm=None,
-    upper_limit_common=None,
-    bayesephem=False,
-    be_type="orbel",
-    wideband=False,
-    dm_var=False,
-    dm_type="gp",
-    dm_psd="powerlaw",
-    dm_annual=False,
-    white_vary=False,
-    gequad=False,
-    dm_chrom=False,
-    dmchrom_psd="powerlaw",
-    dmchrom_idx=4,
-    red_var=True,
-    red_select=None,
-    red_breakflat=False,
-    red_breakflat_fq=None,
-    coefficients=False,
-    pshift=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 2A from the analysis paper:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1.Common red noise modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        2. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum']. 'powerlaw' is default
-        value.
-    :param tm_var: explicitly vary the timing model parameters
-    :param tm_linear: vary the timing model in the linear approximation
-    :param tm_param_list: an explicit list of timing model parameters to vary
-    :param ltm_exclude_list: a list of parameters that will be excluded from being varied linearly
-        if exlude is True; if exclude is False they are the only parameters to include in the linear model
-    :param exclude: bool, whether to include or exlude parameters given in ltm_exclude_list
-    :param tm_param_dict: a nested dictionary of parameters to vary in the model and their user defined values and priors
-    :param tm_prior: prior type on varied timing model parameters {'Uniform','bounded-normal'}
-    :param nltm_plus_ltm: boolean to switch combined non-linear + linear timing models on, only works for tm_var True
-    :param noisedict:
-        Dictionary of pulsar noise properties. Can provide manually,
-        or the code will attempt to find it.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    :param be_type:
-        orbel, orbel-v2, setIII
-    :param wideband:
-        Use wideband par and tim files. Ignore ECORR. Set to False by default.
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-    gp_priors = [upper_limit_red, upper_limit_dm, upper_limit_common]
-    if all(ii is None for ii in gp_priors):
-        amp_prior_red = amp_prior
-        amp_prior_dm = amp_prior
-        amp_prior_common = amp_prior
-    else:
-        amp_prior_red = "uniform" if upper_limit_red else "log-uniform"
-        amp_prior_dm = "uniform" if upper_limit_dm else "log-uniform"
-        amp_prior_common = "uniform" if upper_limit_common else "log-uniform"
-
-    # timing model
-    if not tm_var:
-        s = gp_signals.TimingModel(
-            use_svd=tm_svd, normed=tm_norm, coefficients=coefficients
-        )
-    else:
-        # create new attribute for enterprise pulsar object
-        for p in psrs:
-            p.tm_params_orig = OrderedDict.fromkeys(p.t2pulsar.pars())
-            for key in p.tm_params_orig:
-                p.tm_params_orig[key] = (p.t2pulsar[key].val, p.t2pulsar[key].err)
-        if tm_linear and not nltm_plus_ltm:
-            s = gp_signals.TimingModel(
-            use_svd=tm_svd, normed=tm_norm, coefficients=coefficients
-            )
-        elif not tm_linear and not nltm_plus_ltm:
-            s = timing_block(
-                tm_param_list=tm_param_list,
-                prior_type=tm_prior,
-                prior_sigma=2.0,
-                prior_lower_bound=-3.0,
-                prior_upper_bound=3.0,
-                tm_param_dict=tm_param_dict,
-            )
-        elif not tm_linear and nltm_plus_ltm:
-            #Don't need this catch!
-            if len(ltm_exclude_list) == 0:
-                ltm_exclude_list=tm_param_list
-            psrs2 = [filter_Mmat(psr,ltm_exclude_list=ltm_exclude_list,exclude=exclude) for psr in psrs]
-            s = gp_signals.TimingModel(use_svd=tm_svd, normed=tm_norm, coefficients=coefficients)
-            s += timing_block(
-                tm_param_list=tm_param_list,
-                prior_type=tm_prior,
-                prior_sigma=2.0,
-                prior_lower_bound=-3.0,
-                prior_upper_bound=3.0,
-                tm_param_dict=tm_param_dict,
-            )
-        else:
-            pass
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    if logfreq:
-        fmin = 10.0
-        modes, wgts = model_utils.linBinning(
-            Tspan, nmodes_log, 1.0 / fmin / Tspan, common_components, nmodes_log
-        )
-        wgts = wgts ** 2.0
-
-    if red_var:
-        # red noise
-        s += red_noise_block(
-            psd=red_psd,
-            prior=amp_prior_red,
-            Tspan=Tspan,
-            components=red_components,
-            modes=modes,
-            wgts=wgts,
-            coefficients=coefficients,
-            select=red_select,
-            break_flat=red_breakflat,
-            break_flat_fq=red_breakflat_fq,
-        )
-
-    if common_var:
-        # common red noise block
-        if orf is None:
-            s += common_red_noise_block(
-                psd=common_psd,
-                prior=amp_prior_common,
-                Tspan=Tspan,
-                components=common_components,
-                coefficients=coefficients,
-                gamma_val=gamma_common,
-                name="gw",
-            )
-        elif orf == "hd":
-            s += common_red_noise_block(
-                psd=common_psd,
-                prior=amp_prior_common,
-                Tspan=Tspan,
-                components=common_components,
-                coefficients=coefficients,
-                gamma_val=gamma_common,
-                orf="hd",
-                name="gw",
-            )
-
-    # DM variations
-    if dm_var:
-        if dm_type == "gp":
-            s += dm_noise_block(
-                gp_kernel="diag",
-                psd=dm_psd,
-                prior=amp_prior_dm,
-                components=dm_components,
-                gamma_val=None,
-                coefficients=coefficients,
-            )
-        if dm_annual:
-            s += chrom.dm_annual_signal()
-        if dm_chrom:
-            s += chromatic_noise_block(
-                psd=dmchrom_psd,
-                idx=dmchrom_idx,
-                name="chromatic",
-                components=dm_components,
-                coefficients=coefficients,
-            )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(
-            use_epoch_toas=True, model=be_type
-        )
-    fixed_bayesephem = False
-    if fixed_bayesephem:
-        #TODO: Allow for fixed ephem params in enterprise.deterministic_signals
-        #s += deterministic_signals.PhysicalEphemerisSignal(
-        #    use_epoch_toas=True, model=be_type
-        #)
-        frame_drift_rate = parameter.Uniform(-1e-9, 1e-9)("frame_drift_rate")
-
-        d_jupiter_mass = parameter.Normal(0, 1.54976690e-11)("d_jupiter_mass")
-
-        d_saturn_mass = parameter.Normal(0, 8.17306184e-12)("d_saturn_mass")
-
-        d_uranus_mass = parameter.Normal(0, 5.71923361e-11)("d_uranus_mass")
-
-        d_neptune_mass = parameter.Normal(0, 7.96103855e-11)("d_neptune_mass")
-
-        jup_orb_elements = parameter.Uniform(-0.05, 0.05, size=6)("jup_orb_elements")
-
-        sat_orb_elements = parameter.Uniform(-0.5, 0.5, size=6)("sat_orb_elements")
-
-        # note: default prior for dynamical model is Uniform(-1e-4, 1e-4)
-        #       for each element.
-
-        times, jup_orbit, sat_orbit = utils.get_planet_orbital_elements(model)
-        wf = utils.physical_ephem_delay(
-            frame_drift_rate=frame_drift_rate,
-            d_jupiter_mass=d_jupiter_mass,
-            d_saturn_mass=d_saturn_mass,
-            d_uranus_mass=d_uranus_mass,
-            d_neptune_mass=d_neptune_mass,
-            jup_orb_elements=jup_orb_elements,
-            sat_orb_elements=sat_orb_elements,
-            times=times,
-            jup_orbit=jup_orbit,
-            sat_orbit=sat_orbit,
-        )
-
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    if tm_var and not tm_linear and nltm_plus_ltm:
-        for p in psrs2:
-            if "NANOGrav" in p.flags["pta"] and not wideband:
-                s2 = s + white_noise_block(vary=white_vary, inc_ecorr=True)
-                if gequad:
-                    s2 += white_signals.EquadNoise(
-                        log10_equad=parameter.Uniform(-8.5, -5),
-                        selection=selections.Selection(selections.no_selection),
-                        name="gequad",
-                    )
-                if "1713" in p.name and dm_var:
-                    tmin = p.toas.min() / 86400
-                    tmax = p.toas.max() / 86400
-                    s3 = s2 + chrom.dm_exponential_dip(
-                        tmin=tmin, tmax=tmax, idx=2, sign=False, name="dmexp"
-                    )
-                    models.append(s3(p))
-                else:
-                    models.append(s2(p))
-            else:
-                s4 = s + white_noise_block(vary=white_vary, inc_ecorr=False)
-                if gequad:
-                    s4 += white_signals.EquadNoise(
-                        log10_equad=parameter.Uniform(-8.5, -5),
-                        selection=selections.Selection(selections.no_selection),
-                        name="gequad",
-                    )
-                if "1713" in p.name and dm_var:
-                    tmin = p.toas.min() / 86400
-                    tmax = p.toas.max() / 86400
-                    s5 = s4 + chrom.dm_exponential_dip(
-                        tmin=tmin, tmax=tmax, idx=2, sign=False, name="dmexp"
-                    )
-                    models.append(s5(p))
-                else:
-                    models.append(s4(p))
-    else:
-        for p in psrs:
-            if "NANOGrav" in p.flags["pta"] and not wideband:
-                s2 = s + white_noise_block(vary=white_vary, inc_ecorr=True)
-                if gequad:
-                    s2 += white_signals.EquadNoise(
-                        log10_equad=parameter.Uniform(-8.5, -5),
-                        selection=selections.Selection(selections.no_selection),
-                        name="gequad",
-                    )
-                if "1713" in p.name and dm_var:
-                    tmin = p.toas.min() / 86400
-                    tmax = p.toas.max() / 86400
-                    s3 = s2 + chrom.dm_exponential_dip(
-                        tmin=tmin, tmax=tmax, idx=2, sign=False, name="dmexp"
-                    )
-                    models.append(s3(p))
-                else:
-                    models.append(s2(p))
-            else:
-                s4 = s + white_noise_block(vary=white_vary, inc_ecorr=False)
-                if gequad:
-                    s4 += white_signals.EquadNoise(
-                        log10_equad=parameter.Uniform(-8.5, -5),
-                        selection=selections.Selection(selections.no_selection),
-                        name="gequad",
-                    )
-                if "1713" in p.name and dm_var:
-                    tmin = p.toas.min() / 86400
-                    tmax = p.toas.max() / 86400
-                    s5 = s4 + chrom.dm_exponential_dip(
-                        tmin=tmin, tmax=tmax, idx=2, sign=False, name="dmexp"
-                    )
-                    models.append(s5(p))
-                else:
-                    models.append(s4(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if not white_vary:
-        if noisedict is None:
-            print("No noise dictionary provided!...")
-        else:
-            noisedict = noisedict
-            pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_2b(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    bayesephem=False,
-    wideband=False,
-    pshift=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 2B from the analysis paper:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1. Dipole spatially correlated signal modeled with PSD.
-        Default PSD is powerlaw. Available options
-        ['powerlaw', 'turnover', 'spectrum']
-        2. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum']. 'powerlaw' is default
-        value.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(prior=amp_prior, Tspan=Tspan, components=components)
-
-    # dipole
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="dipole",
-        name="dipole",
-        pshift=pshift,
-    )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_2c(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    bayesephem=False,
-    wideband=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 2C from the analysis paper:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1. Dipole spatially correlated signal modeled with PSD.
-        Default PSD is powerlaw. Available options
-        ['powerlaw', 'turnover', 'spectrum']
-        2. Monopole spatially correlated signal modeled with PSD.
-        Default PSD is powerlaw. Available options
-        ['powerlaw', 'turnover', 'spectrum']
-        3. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum']. 'powerlaw' is default
-        value.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(prior=amp_prior, Tspan=Tspan, components=components)
-
-    # dipole
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="dipole",
-        name="dipole",
-    )
-
-    # monopole
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="monopole",
-        name="monopole",
-    )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_2d(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    bayesephem=False,
-    wideband=False,
-    pshift=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 2D from the analysis paper:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1. Monopole spatially correlated signal modeled with PSD.
-        Default PSD is powerlaw. Available options
-        ['powerlaw', 'turnover', 'spectrum']
-        2. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum']. 'powerlaw' is default
-        value.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(prior=amp_prior, Tspan=Tspan, components=components)
-
-    # monopole
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="monopole",
-        name="monopole",
-        pshift=pshift,
-    )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_3a(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    bayesephem=False,
-    be_type="orbel",
-    wideband=False,
-    correlationsonly=False,
-    pshift=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 3A from the analysis paper:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1. GWB with HD correlations modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        2. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum'] 'powerlaw' is default
-        value.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    :param be_type:
-        orbel, orbel-v2, setIII
-    :param correlationsonly:
-        Give infinite power (well, 1e40) to pulsar red noise, effectively
-        canceling out also GW diagonal terms
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(
-        prior="infinitepower" if correlationsonly else amp_prior,
-        Tspan=Tspan,
-        components=components,
-    )
-
-    # common red noise block
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="hd",
-        name="gw",
-        pshift=pshift,
-    )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(
-            use_epoch_toas=True, model=be_type
-        )
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_3b(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    bayesephem=False,
-    wideband=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 3B from the analysis paper:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1. GWB with HD correlations modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        2. Dipole signal modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        3. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum'] 'powerlaw' is default
-        value.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(prior=amp_prior, Tspan=Tspan, components=components)
-
-    # common red noise block
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="hd",
-        name="gw",
-    )
-
-    # dipole
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="dipole",
-        name="dipole",
-    )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_3c(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    bayesephem=False,
-    wideband=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 3C from the analysis paper:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1. GWB with HD correlations modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        2. Dipole signal modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        3. Monopole signal modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        4. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum'] 'powerlaw' is default
-        value.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(prior=amp_prior, Tspan=Tspan, components=components)
-
-    # common red noise block
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="hd",
-        name="gw",
-    )
-
-    # dipole
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="dipole",
-        name="dipole",
-    )
-
-    # monopole
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="monopole",
-        name="monopole",
-    )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_3d(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    bayesephem=False,
-    wideband=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 3D from the analysis paper:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1. GWB with HD correlations modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        2. Monopole signal modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        3. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum'] 'powerlaw' is default
-        value.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(prior=amp_prior, Tspan=Tspan, components=components)
-
-    # common red noise block
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="hd",
-        name="gw",
-    )
-
-    # monopole
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        orf="monopole",
-        name="monopole",
-    )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_2a_drop_be(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    wideband=False,
-    k_threshold=0.5,
-    pshift=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 2A from the analysis paper:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1.Common red noise modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        2. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum']. 'powerlaw' is default
-        value.
-    :param noisedict:
-        Dictionary of pulsar noise properties. Can provide manually,
-        or the code will attempt to find it.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param wideband:
-        Use wideband par and tim files. Ignore ECORR. Set to False by default.
-    :param k_threshold:
-        Define threshold for dropout parameter 'k'.
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(prior=amp_prior, Tspan=Tspan, components=components)
-
-    # common red noise block
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        name="gw",
-        pshift=pshift,
-    )
-
-    # ephemeris model
-    s += do.Dropout_PhysicalEphemerisSignal(
-        use_epoch_toas=True, k_threshold=k_threshold
-    )
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_2a_drop_crn(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    bayesephem=False,
-    wideband=False,
-    k_threshold=0.5,
-    pshift=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 2A from the analysis paper:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-
-    global:
-        1.Common red noise modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        2. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum']. 'powerlaw' is default
-        value.
-    :param noisedict:
-        Dictionary of pulsar noise properties. Can provide manually,
-        or the code will attempt to find it.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    :param wideband:
-        Use wideband par and tim files. Ignore ECORR. Set to False by default.
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # red noise
-    s = red_noise_block(prior=amp_prior, Tspan=Tspan, components=components)
-
-    # common red noise block
-    amp_name = "{}_log10_A".format("gw")
-    if amp_prior == "uniform":
-        log10_Agw = parameter.LinearExp(-18, -11)(amp_name)
-    elif amp_prior == "log-uniform" and gamma_common is not None:
-        if np.abs(gamma_common - 4.33) < 0.1:
-            log10_Agw = parameter.Uniform(-18, -14)(amp_name)
-        else:
-            log10_Agw = parameter.Uniform(-18, -11)(amp_name)
-    else:
-        log10_Agw = parameter.Uniform(-18, -11)(amp_name)
-
-    gam_name = "{}_gamma".format("gw")
-    if gamma_common is not None:
-        gamma_gw = parameter.Constant(gamma_common)(gam_name)
-    else:
-        gamma_gw = parameter.Uniform(0, 7)(gam_name)
-
-    k_drop = parameter.Uniform(0.0, 1.0)  # per-pulsar
-
-    drop_pl = do.dropout_powerlaw(
-        log10_A=log10_Agw, gamma=gamma_gw, k_drop=k_drop, k_threshold=k_threshold
-    )
-    crn = gp_signals.FourierBasisGP(
-        drop_pl, components=components, Tspan=Tspan, name="gw", pshift=pshift
-    )
-    s += crn
-
-    # ephemeris model
-    s += do.Dropout_PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-# Does not yet work with IPTA datasets due to white-noise modeling issues.
-def model_chromatic(
-    psrs,
-    psd="powerlaw",
-    noisedict=None,
-    components=30,
-    gamma_common=None,
-    upper_limit=False,
-    bayesephem=False,
-    wideband=False,
-    pshift=False,
-    idx=4,
-    chromatic_psd="powerlaw",
-    c_psrs=["J1713+0747"],
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with model 2A from the analysis paper + additional
-    chromatic noise for given pulsars
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-        6. Chromatic noise for given pulsar list
-
-    global:
-        1.Common red noise modeled with user defined PSD with
-        30 sampling frequencies. Available PSDs are
-        ['powerlaw', 'turnover' 'spectrum']
-        2. Optional physical ephemeris modeling.
-
-    :param psd:
-        PSD to use for common red noise signal. Available options
-        are ['powerlaw', 'turnover' 'spectrum']. 'powerlaw' is default
-        value.
-    :param noisedict:
-        Dictionary of pulsar noise properties. Can provide manually,
-        or the code will attempt to find it.
-    :param gamma_common:
-        Fixed common red process spectral index value. By default we
-        vary the spectral index over the range [0, 7].
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    :param wideband:
-        Use wideband par and tim files. Ignore ECORR. Set to False by default.
-    :param idx:
-        Index of chromatic process (i.e DM is 2, scattering would be 4). If
-        set to `vary` then will vary from 0 - 6 (This will be VERY slow!)
-    :param chromatic_psd:
-        PSD to use for chromatic noise. Available options
-        are ['powerlaw', 'turnover' 'spectrum']. 'powerlaw' is default
-        value.
-    :param c_psrs:
-        List of pulsars to use chromatic noise. 'all' will use all pulsars
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    Tspan = model_utils.get_tspan(psrs)
-
-    # white noise
-    s = white_noise_block(vary=False, wideband=wideband)
-
-    # red noise
-    s += red_noise_block(prior=amp_prior, Tspan=Tspan, components=components)
-
-    # common red noise block
-    s += common_red_noise_block(
-        psd=psd,
-        prior=amp_prior,
-        Tspan=Tspan,
-        components=components,
-        gamma_val=gamma_common,
-        name="gw",
-        pshift=pshift,
-    )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # chromatic noise
-    sc = chromatic_noise_block(psd=chromatic_psd, idx=idx)
-    if c_psrs == "all":
-        s += sc
-        models = [s(psr) for psr in psrs]
-    elif len(c_psrs) > 0:
-        models = []
-        for psr in psrs:
-            if psr.name in c_psrs:
-                print("Adding chromatic model to PSR {}".format(psr.name))
-                snew = s + sc
-                models.append(snew(psr))
-            else:
-                models.append(s(psr))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_bwm(
-    psrs,
-    noisedict=None,
-    tm_svd=False,
-    Tmin_bwm=None,
-    Tmax_bwm=None,
-    skyloc=None,
-    red_psd="powerlaw",
-    components=30,
-    dm_var=False,
-    dm_psd="powerlaw",
-    dm_annual=False,
-    upper_limit=False,
-    bayesephem=False,
-    wideband=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with BWM model:
-
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system (if NG channelized)
-        4. Red noise modeled by a specified psd
-        5. Linear timing model.
-        6. Optional DM-variation modeling
-    global:
-        1. Deterministic GW burst with memory signal.
-        2. Optional physical ephemeris modeling.
-
-    :param psrs:
-        list of enterprise.Pulsar objects for PTA
-    :param noisedict:
-        Dictionary of pulsar noise properties for fixed white noise.
-        Can provide manually, or the code will attempt to find it.
-    :param tm_svd:
-        boolean for svd-stabilised timing model design matrix
-    :param Tmin_bwm:
-        Min time to search for BWM (MJD). If omitted, uses first TOA.
-    :param Tmax_bwm:
-        Max time to search for BWM (MJD). If omitted, uses last TOA.
-    :param skyloc:
-        Fixed sky location of BWM signal search as [cos(theta), phi].
-        Search over sky location if ``None`` given.
-    :param red_psd:
-        PSD to use for per pulsar red noise. Available options
-        are ['powerlaw', 'turnover', tprocess, 'spectrum'].
-    :param components:
-        number of modes in Fourier domain processes (red noise, DM
-        variations, etc)
-    :param dm_var:
-        include gaussian process DM variations
-    :param dm_psd:
-        power-spectral density for gp DM variations
-    :param dm_annual:
-        include a yearly period DM variation
-    :param upper_limit:
-        Perform upper limit on BWM amplitude. By default this is
-        set to False for a 'detection' run.
-    :param bayesephem:
-        Include BayesEphem model.
-    :return: instantiated enterprise.PTA object
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set frequency sampling
-    tmin = np.min([p.toas.min() for p in psrs])
-    tmax = np.max([p.toas.max() for p in psrs])
-    Tspan = tmax - tmin
-
-    if Tmin_bwm is None:
-        Tmin_bwm = tmin / const.day
-    if Tmax_bwm is None:
-        Tmax_bwm = tmax / const.day
-
-    # red noise
-    s = red_noise_block(
-        prior=amp_prior, psd=red_psd, Tspan=Tspan, components=components
-    )
-
-    # DM variations
-    if dm_var:
-        s += dm_noise_block(
-            psd=dm_psd, prior=amp_prior, components=components, gamma_val=None
-        )
-        if dm_annual:
-            s += chrom.dm_annual_signal()
-
-        # DM exponential dip for J1713's DM event
-        dmexp = chrom.dm_exponential_dip(tmin=54500, tmax=54900)
-
-    # GW BWM signal block
-    s += deterministic.bwm_block(
-        Tmin_bwm, Tmax_bwm, amp_prior=amp_prior, skyloc=skyloc, name="bwm"
-    )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel(use_svd=tm_svd)
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True)
-            if dm_var and "J1713+0747" == p.name:
-                s2 += dmexp
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            if dm_var and "J1713+0747" == p.name:
-                s3 += dmexp
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
-
-
-def model_cw(
-    psrs,
-    upper_limit=False,
-    noisedict=None,
-    rn_psd="powerlaw",
-    components=30,
-    bayesephem=False,
-    skyloc=None,
-    log10_F=None,
-    ecc=False,
-    psrTerm=False,
-    wideband=False,
-):
-    """
-    Reads in list of enterprise Pulsar instance and returns a PTA
-    instantiated with CW model:
-    per pulsar:
-        1. fixed EFAC per backend/receiver system
-        2. fixed EQUAD per backend/receiver system
-        3. fixed ECORR per backend/receiver system
-        4. Red noise modeled as a power-law with 30 sampling frequencies
-        5. Linear timing model.
-    global:
-        1. Deterministic CW signal.
-        2. Optional physical ephemeris modeling.
-    :param upper_limit:
-        Perform upper limit on common red noise amplitude. By default
-        this is set to False. Note that when perfoming upper limits it
-        is recommended that the spectral index also be fixed to a specific
-        value.
-    :param rn_psd:
-        psd to use in red_noise_block()
-    :param bayesephem:
-        Include BayesEphem model. Set to False by default
-    :param skyloc:
-        Fixed sky location of CW signal search as [cos(theta), phi].
-        Search over sky location if ``None`` given.
-    :param log10_F:
-        Fixed frequency of CW signal search.
-        Search over frequency if ``None`` given.
-    :param ecc:
-        boolean or float
-        if boolean: include/exclude eccentricity in search
-        if float: use fixed eccentricity with eccentric model
-    :psrTerm:
-        boolean, include/exclude pulsar term in search
-    """
-
-    amp_prior = "uniform" if upper_limit else "log-uniform"
-
-    # find the maximum time span to set GW frequency sampling
-    tmin = np.min([p.toas.min() for p in psrs])
-    tmax = np.max([p.toas.max() for p in psrs])
-    Tspan = tmax - tmin
-
-    # red noise
-    s = red_noise_block(prior=amp_prior, psd=rn_psd, Tspan=Tspan, components=components)
-
-    # GW CW signal block
-    if not ecc:
-        s += deterministic.cw_block_circ(
-            amp_prior=amp_prior,
-            skyloc=skyloc,
-            log10_fgw=log10_F,
-            psrTerm=psrTerm,
-            tref=tmin,
-            name="cw",
-        )
-    else:
-        if type(ecc) is not float:
-            ecc = None
-        s += deterministic.cw_block_ecc(
-            amp_prior=amp_prior,
-            skyloc=skyloc,
-            log10_F=log10_F,
-            ecc=ecc,
-            psrTerm=psrTerm,
-            tref=tmin,
-            name="cw",
-        )
-
-    # ephemeris model
-    if bayesephem:
-        s += deterministic_signals.PhysicalEphemerisSignal(use_epoch_toas=True)
-
-    # timing model
-    s += gp_signals.TimingModel()
-
-    # adding white-noise, and acting on psr objects
-    models = []
-    for p in psrs:
-        if "NANOGrav" in p.flags["pta"] and not wideband:
-            s2 = s + white_noise_block(vary=False, inc_ecorr=True, gp_ecorr=True)
-            models.append(s2(p))
-        else:
-            s3 = s + white_noise_block(vary=False, inc_ecorr=False)
-            models.append(s3(p))
-
-    # set up PTA
-    pta = signal_base.PTA(models)
-
-    # set white noise parameters
-    if noisedict is None:
-        print("No noise dictionary provided!...")
-    else:
-        noisedict = noisedict
-        pta.set_default_params(noisedict)
-
-    return pta
